make.sh: Include clean script + minor hack I decided to throw in the clean script which allows a developer to run make.sh after cloning without needing to run clean.sh. This is very minor but is more convenient for me.
I also define the cmake toolchain twice which gets around a strange but that prevents the compile from completing. Before this change you could run make.sh twice to get a completed compilation, but this removes the need for that. I would like to figure out why this is needed, but it's not really high on my priority list.